Improve Product Index page refresh on deletion. DRY code
