+++ app/views/carts/_cart_items.html.erb
@@ -13,10 +13,5 @@
       <td><%= link_to(t('buttons.remove_it'), cart_item_path(cart_item), data: {turbo_method: :delete})  %></td>
     </tbody>
   <% end %>
-  <% @cart.cart_items.map do |cart_item| %>
-    <% qt = cart_item.quantity %>
-    <% pr = cart_item.product.price %>
-    <% @sum += qt * pr %>
-  <% end %>
 </div>
