+++ app/views/products/index.html.erb
@@ -1,4 +1,7 @@
-<h1> <%= t('.title') %> </h1>
+<div>
+  <h1><%= @product_category.name %></h1>
+  <h3><%= link_to(t('.add_new_product'), new_product_path) if policy(Product).index? %></h3>
+</div>
 <% instance_variable_names.inspect %>
 <div>
   <%= form_for products_path, data: { turbo_frame: :results }, method: :get do %>
@@ -26,7 +29,6 @@
           <%= search_form.radio_button :sort, 'price asc', oninput: 'this.form.requestSubmit()' %>
           <%= search_form.radio_button :sort, 'price desc', oninput: 'this.form.requestSubmit()' %>
         </div>
-
       </div>
       <%= search_form.hidden_field :product_category_id, value: @product_category.id %>
     <% end %>
@@ -43,7 +45,3 @@
   <%= turbo_frame_tag :pagination, loading: :lazy, src: products_path(format: :turbo_stream, search: search_params) %>
 <% end %>
 
-
-<div>
-  <div><%= link_to(t('buttons.new'), new_product_path) if policy(Product).index?  %></div>
-</div>
